Penambahan upload multiple bukti close

+++ app/Http/Controllers/LUP/LUPFileController.php
@@ -59,6 +59,55 @@ class LUPFileController extends Controller
            
     }
 
+    public function uploadevidence(Request $request,$id)
+    {
+        
+        //get data lupparent
+        $lup = lupparent::find($id);
+        $paths = DB::table('iccsfilepaths')
+        ->where('description','Lampiran LUP')
+        ->first()->filepath;        
+        $id= Crypt::encryptString($lup->id);
+
+        //get code attachment
+        $lastid = lupFile::where('code',$lup->code)->max('nofile');
+        $lastno = intval(substr($lastid,-2));
+        
+        if($lastid==0 or $lastid==NULL){
+            $newid = 1;
+        }else{
+            $newid = abs($lastno+1);
+        }        
+        $code = $lup->code.'-ATT-'. sprintf("%02s", $newid);        
+        
+        
+        $validatedData = $request->validate([
+            'attachment_file' => 'required|mimes:xls,xlsx,pdf,jpg,png,jpeg,doc,docx,msg,eml|max:1000',
+    
+           ]);
+    
+           $name = $request->file('attachment_file')->getClientOriginalName();
+           $ext = pathinfo($name, PATHINFO_EXTENSION);
+           $path = $request->file('attachment_file')->storeAs($paths,$code.'.'.$ext);
+    
+           
+            $save = lupFile::create([
+            'code' => $lup->code,    
+            'nofile'=>$code,
+            'org_file_name'=>$name,
+            'document_name'=>$request->modaltxtadddocname, 
+            'uploader'=>Auth::user()->username,
+            'date_upload'=> \Carbon\Carbon::now(),  
+            'file_path' => $path,
+            'is_evidence'=>true,
+            'action'=>$request->referaction,            
+        ]); 
+           
+        return redirect('/lup/'.$id.'/edit')->with('info', 'File Has been uploaded successfully');       
+           
+    }
+
+
     
     //download attachment lup
     public function download($id)
@@ -121,8 +170,7 @@ class LUPFileController extends Controller
         if($file->wasChanged()==True){
             auditlups($file,Auth::user()->username,'Change Attachment',$file->code,
                 'lupfiles','document_name',$oldname.' ; '.$oldfile,$file->document_name.' ; '.$file->org_file_name );
-        }
-         
+        }       
                                           
         return redirect('/lup/'.$id.'/edit')->with('info', 'File Has been uploaded successfully');       
            
@@ -133,6 +181,8 @@ class LUPFileController extends Controller
     {
         $decrypted = Crypt::decryptString($id);
         $file = lupFile::find($decrypted);
+        $lup=$file->lupparent;
+        $this->authorize('update',$lup);
         auditlups($file,Auth::user()->username,'Delete Attachment',$file->code,
                 'lupfiles','',$file->makeHidden(['id', 'deleted_at','file_path','longtext1','longtext2','longtext3','date1','date2','date3']),'');
         $file->delete();
